Removed several now unused classes and files. Further refining of new PyRep dumping.
