+++ src/eve-common/cache/CachedObjectMgr.h
@@ -200,51 +200,4 @@ public:
         //copy of the version tuple in rret
 };
 
-//run through the rep, concatenating all the strings together and noting if
-//there are a no non-string types in the rep (lists and tuples are OK)
-class StringCollapseVisitor
-: public PyVisitor
-{
-public:
-    std::string result;
-
-    bool VisitInteger( const PyInt* rep )
-    {
-        if( !result.empty() )
-            result += ".";
-
-        std::stringstream ss;
-        ss << rep->value();
-        result += ss.str();
-
-        return true;
-    }
-    bool VisitLong( const PyLong* rep ) { return false; }
-    bool VisitReal( const PyFloat* rep ) { return false; }
-    bool VisitBoolean( const PyBool* rep ) { return false; }
-    bool VisitNone( const PyNone* rep ) { return false; }
-    bool VisitBuffer( const PyBuffer* rep ) { return false; }
-    bool VisitString( const PyString* rep )
-    {
-        if( !result.empty() )
-            result += ".";
-        result += rep->content();
-
-        return true;
-    }
-    bool VisitWString( const PyWString* rep ) { return false; }
-    bool VisitToken( const PyToken* rep ) { return false; }
-
-    bool VisitDict( const PyDict* rep ) { return false; }
-
-    bool VisitObject( const PyObject* rep ) { return false; }
-    bool VisitObjectEx( const PyObjectEx* rep ) { return false; }
-
-    bool VisitPackedRow( const PyPackedRow* rep ) { return false; }
-
-    bool VisitSubStruct( const PySubStruct* rep ) { return false; }
-    bool VisitSubStream( const PySubStream* rep ) { return false; }
-    bool VisitChecksumedStream( const PyChecksumedStream* rep ) { return false; }
-};
-
 #endif
