+++ src/eve-common/network/EVESession.cpp
@@ -31,7 +31,6 @@
 #include "packets/Crypto.h"
 #include "python/PyRep.h"
 #include "python/PyPacket.h"
-#include "python/PyDumpVisitor.h"
 #include "EVEVersion.h"
 
 EVEClientSession::EVEClientSession(EVETCPConnection** n)
@@ -90,7 +89,7 @@ void EVEClientSession::fastQueuePacket(PyPacket** p)
         std::string pfx = getLogPrefix(CLIENT__OUT_ALL);
         std::ostringstream ss;
         (*p)->Dump(ss, pfx);
-        _log(CLIENT__OUT_ALL, ss.str().c_str());
+        outputLogMsg(CLIENT__OUT_ALL, ss.str().c_str());
     }
 
     PyRep* r = (*p)->Encode();
