+++ src/eve-server/Client.cpp
@@ -126,10 +126,11 @@ bool Client::ProcessNet()
     {
         if(is_log_enabled(CLIENT__IN_ALL))
         {
+            _log(CLIENT__IN_ALL, "Received packet:");
             std::string pfx = getLogPrefix(CLIENT__IN_ALL);
             std::ostringstream ss;
             p->Dump(ss, pfx);
-            _log(CLIENT__IN_ALL, "Received packet:\n%s", ss.str());
+            outputLogMsg(CLIENT__IN_ALL, ss.str().c_str());
         }
 
         try
